+++ src/components/Button/Button.styles.ts
@@ -1,12 +1,5 @@
 import styled from "@emotion/styled";
 
-interface ButtonProps {
-  text: string;
-  colorType: number;
-  isDisabled: boolean;
-  onClick?: () => void;
-}
-
 interface ButtonWrapperProps {
   colorType: number;
 }
@@ -115,20 +108,3 @@ export const ButtonWrapper = styled.div<ButtonWrapperProps>`
     }};
   }
 `;
-
-const Button: React.FC<ButtonProps> = ({
-  text,
-  colorType,
-  isDisabled,
-  onClick,
-}) => {
-  return (
-    <ButtonWrapper colorType={colorType} onClick={onClick}>
-      <button className={`btn ${isDisabled ? "disabled" : "active"}`}>
-        {text}
-      </button>
-    </ButtonWrapper>
-  );
-};
-
-export default Button;
